pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "edge-tts>=7.0.0",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "numpy>=2.2.3",
    "pytz>=2025.1",
    "requests>=2.32.3",
    "speechrecognition>=3.14.1",
]
